pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-bootstrap5==24.3",
    "django-filter>=25.1",
    "django-extensions>=4.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rollbar>=1.3.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager", "config"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-p no:pytest_dotenv -p no:pytest-dotenv -p no:pytest_env -p no:pytest-env"
pytest_plugins = ["code.pytest_plugin_blocker"]
